requirements.txt
pystray>=0.19.5
Pillow>=10.0.0
requests>=2.31.0
plyer>=2.1.0
# Для HTML-окна (левый клик → дашборд).
# Если не нужно — можно не ставить, скрипт автоматически переключится на tkinter.
pywebview>=5.0.0
